move controller.NewController to controller/clusterapi.NewController
